Fixed some things and added in Firebase logout on all pages, as well as have all enterable information on every page savable but no link to firebase.
Fix 1: Made a change to the AndroidManifest.xml that had both the home page and login activities as the start page, made it the login page only. Seems to still open up on the homepage sometimes.

Fix 2: Added Firebase sign out from every page in the app.

Functionality: Added a drop down menu for the priority setting for a user created category. This information is saved but not to Firebase. All information is saved now I believe, but none of it besides login/create account stuff is saved to the user account on Firebase.